Moved all tests to same file for ease of accessibility
