+++ azblob/parsing_urls.go
@@ -1,7 +1,7 @@
 package azblob
 
 import (
-	"log"
+	"errors"
 	"net"
 	"net/url"
 	"strings"
@@ -136,7 +136,7 @@ func (up BlobURLParts) URL() url.URL {
 
 	// Check: Both snapshot and version id cannot be present in the request URL.
 	if up.Snapshot != "" && up.VersionID != "" {
-		log.Fatal("Snapshot and versioning cannot be enabled simultaneously.")
+		errors.New("Snapshot and versioning cannot be enabled simultaneously")
 	}
 
 	//If no snapshot is initially provided, fill it in from the SAS query properties to help the user
