+++ azblob/zt_url_blob_test.go
@@ -1971,191 +1971,3 @@ func (s *aztestsSuite) TestDownloadBlockBlobUnexpectedEOF(c *chk.C) {
 	c.Assert(buf, chk.DeepEquals, []byte(blockBlobDefaultData))
 }
 
-func (s *aztestsSuite) TestSetBlobMetadataReturnsVID(c *chk.C) {
-	bsu := getBSU()
-	containerURL, _ := createNewContainer(c, bsu)
-	defer deleteContainer(c, containerURL)
-	blobURL, blobName := createNewBlockBlob(c, containerURL)
-	metadata := Metadata{"test_key_1": "test_value_1", "test_key_2": "2019"}
-	resp, err := blobURL.SetMetadata(ctx, metadata, BlobAccessConditions{})
-	c.Assert(err, chk.IsNil)
-	c.Assert(resp.VersionID(), chk.NotNil)
-
-	listBlobResp, err := containerURL.ListBlobsFlatSegment(ctx, Marker{}, ListBlobsSegmentOptions{Details: BlobListingDetails{Metadata: true}})
-
-	c.Assert(err, chk.IsNil)
-	c.Assert(listBlobResp.Segment.BlobItems[0].Name, chk.Equals, blobName)
-	c.Assert(listBlobResp.Segment.BlobItems[0].Metadata, chk.HasLen, 2)
-	c.Assert(listBlobResp.Segment.BlobItems[0].Metadata, chk.DeepEquals, metadata)
-}
-
-func (s *aztestsSuite) TestCreateAndDownloadBlobSpecialCharactersWithVID(c *chk.C) {
-	bsu := getBSU()
-	containerURL, _ := createNewContainer(c, bsu)
-	defer deleteContainer(c, containerURL)
-	data := []rune("-._/()$=',~0123456789")
-	for i := 0; i < len(data); i++ {
-		blobName := "abc" + string(data[i])
-		blobURL := containerURL.NewBlockBlobURL(blobName)
-		resp, err := blobURL.Upload(ctx, strings.NewReader(string(data[i])), BlobHTTPHeaders{}, Metadata{}, BlobAccessConditions{})
-		c.Assert(err, chk.IsNil)
-		c.Assert(resp.VersionID(), chk.NotNil)
-
-		dResp, err := blobURL.WithVersionID(resp.VersionID()).Download(ctx, 0, CountToEnd, BlobAccessConditions{}, false)
-		d1, err := ioutil.ReadAll(dResp.Body(RetryReaderOptions{}))
-		c.Assert(err, chk.IsNil)
-		c.Assert(dResp.Version(), chk.Not(chk.Equals), "")
-		c.Assert(string(d1), chk.DeepEquals, string(data[i]))
-		versionId := dResp.r.rawResponse.Header.Get("x-ms-version-id")
-		c.Assert(versionId, chk.NotNil)
-		c.Assert(versionId, chk.Equals, resp.VersionID())
-	}
-}
-
-func (s *aztestsSuite) TestDeleteSpecificBlobVersion(c *chk.C) {
-	bsu := getBSU()
-	containerURL, _ := createNewContainer(c, bsu)
-	defer deleteContainer(c, containerURL)
-	blobURL, _ := getBlockBlobURL(c, containerURL)
-
-	blockBlobUploadResp, err := blobURL.Upload(ctx, bytes.NewReader([]byte("data")), BlobHTTPHeaders{},
-		basicMetadata, BlobAccessConditions{})
-	c.Assert(err, chk.IsNil)
-	c.Assert(blockBlobUploadResp.VersionID(), chk.NotNil)
-	versionId1 := blockBlobUploadResp.VersionID()
-
-	blockBlobUploadResp, err = blobURL.Upload(ctx, bytes.NewReader([]byte("updated_data")), BlobHTTPHeaders{},
-		basicMetadata, BlobAccessConditions{})
-	c.Assert(err, chk.IsNil)
-	c.Assert(blockBlobUploadResp.VersionID(), chk.NotNil)
-
-	listBlobsResp, err := containerURL.ListBlobsFlatSegment(ctx, Marker{}, ListBlobsSegmentOptions{Details: BlobListingDetails{Versions: true}})
-	c.Assert(err, chk.IsNil)
-	c.Assert(listBlobsResp.Segment.BlobItems, chk.HasLen, 2)
-
-	// Deleting previous version snapshot.
-	deleteResp, err := blobURL.WithVersionID(versionId1).Delete(ctx, DeleteSnapshotsOptionNone, BlobAccessConditions{})
-	c.Assert(err, chk.IsNil)
-	c.Assert(deleteResp.StatusCode(), chk.Equals, 202)
-
-	listBlobsResp, err = containerURL.ListBlobsFlatSegment(ctx, Marker{}, ListBlobsSegmentOptions{Details: BlobListingDetails{Versions: true}})
-	c.Assert(err, chk.IsNil)
-	c.Assert(listBlobsResp.Segment.BlobItems, chk.NotNil)
-	if len(listBlobsResp.Segment.BlobItems) == 0 {
-		c.Fail()
-	}
-}
-
-func (s *aztestsSuite) TestDeleteSpecificBlobVersionWithBlobSAS(c *chk.C) {
-	bsu := getBSU()
-	credential, err := getGenericCredential("")
-	if err != nil {
-		c.Fatal(err)
-	}
-	containerURL, containerName := createNewContainer(c, bsu)
-	defer deleteContainer(c, containerURL)
-	blobURL, blobName := getBlockBlobURL(c, containerURL)
-
-	resp, err := blobURL.Upload(ctx, bytes.NewReader([]byte("data")), BlobHTTPHeaders{},
-		basicMetadata, BlobAccessConditions{})
-	c.Assert(err, chk.IsNil)
-	versionId := resp.VersionID()
-	c.Assert(versionId, chk.NotNil)
-
-	resp, err = blobURL.Upload(ctx, bytes.NewReader([]byte("updated_data")), BlobHTTPHeaders{},
-		basicMetadata, BlobAccessConditions{})
-	c.Assert(err, chk.IsNil)
-	c.Assert(resp.VersionID(), chk.NotNil)
-
-	blobParts := NewBlobURLParts(blobURL.URL())
-	blobParts.VersionID = versionId
-	blobParts.SAS, err = BlobSASSignatureValues{
-		Protocol:      SASProtocolHTTPS,
-		ExpiryTime:    time.Now().UTC().Add(1 * time.Hour),
-		ContainerName: containerName,
-		BlobName:      blobName,
-		Permissions:   BlobSASPermissions{Delete: true, DeletePreviousVersion: true}.String(),
-	}.NewSASQueryParameters(credential)
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	sbURL := NewBlockBlobURL(blobParts.URL(), containerURL.client.p)
-	deleteResp, err := sbURL.Delete(ctx, DeleteSnapshotsOptionNone, BlobAccessConditions{})
-	c.Assert(deleteResp, chk.IsNil)
-
-	listBlobResp, err := containerURL.ListBlobsFlatSegment(ctx, Marker{}, ListBlobsSegmentOptions{Details: BlobListingDetails{Versions: true}})
-	c.Assert(err, chk.IsNil)
-	c.Assert(listBlobResp.Version(), chk.Equals, ServiceVersion)
-	for _, blob := range listBlobResp.Segment.BlobItems {
-		c.Assert(blob.VersionID, chk.Not(chk.Equals), versionId)
-	}
-}
-
-func (s *aztestsSuite) TestDownloadSpecificBlobVersion(c *chk.C) {
-	bsu := getBSU()
-	containerURL, _ := createNewContainer(c, bsu)
-	defer deleteContainer(c, containerURL)
-	blobURL, _ := getBlockBlobURL(c, containerURL)
-
-	blockBlobUploadResp, err := blobURL.Upload(ctx, bytes.NewReader([]byte("data")), BlobHTTPHeaders{},
-		basicMetadata, BlobAccessConditions{})
-	c.Assert(err, chk.IsNil)
-	c.Assert(blockBlobUploadResp, chk.NotNil)
-	versionId1 := blockBlobUploadResp.VersionID()
-
-	blockBlobUploadResp, err = blobURL.Upload(ctx, bytes.NewReader([]byte("updated_data")), BlobHTTPHeaders{},
-		basicMetadata, BlobAccessConditions{})
-	c.Assert(err, chk.IsNil)
-	c.Assert(blockBlobUploadResp, chk.NotNil)
-	versionId2 := blockBlobUploadResp.VersionID()
-	c.Assert(blockBlobUploadResp.VersionID(), chk.NotNil)
-
-	// Download previous version of snapshot.
-	blobURL = blobURL.WithVersionID(versionId1)
-	blockBlobDeleteResp, err := blobURL.Download(ctx, 0, CountToEnd, BlobAccessConditions{}, false)
-	c.Assert(err, chk.IsNil)
-	data, err := ioutil.ReadAll(blockBlobDeleteResp.Response().Body)
-	c.Assert(string(data), chk.Equals, "data")
-
-	// Download current version of snapshot.
-	blobURL = blobURL.WithVersionID(versionId2)
-	blockBlobDeleteResp, err = blobURL.Download(ctx, 0, CountToEnd, BlobAccessConditions{}, false)
-	c.Assert(err, chk.IsNil)
-	data, err = ioutil.ReadAll(blockBlobDeleteResp.Response().Body)
-	c.Assert(string(data), chk.Equals, "updated_data")
-}
-
-func (s *aztestsSuite) TestCreateBlobSnapshotReturnsVID(c *chk.C) {
-	bsu := getBSU()
-	containerURL, _ := createNewContainer(c, bsu)
-	defer delContainer(c, containerURL)
-	blobURL := containerURL.NewBlockBlobURL(generateBlobName())
-	uploadResp, err := blobURL.Upload(ctx, bytes.NewReader([]byte("updated_data")), BlobHTTPHeaders{},
-		basicMetadata, BlobAccessConditions{})
-	c.Assert(err, chk.IsNil)
-	c.Assert(uploadResp.VersionID(), chk.NotNil)
-
-	csResp, err := blobURL.CreateSnapshot(ctx, Metadata{}, BlobAccessConditions{})
-	c.Assert(err, chk.IsNil)
-	c.Assert(csResp.VersionID(), chk.NotNil)
-	lbResp, err := containerURL.ListBlobsFlatSegment(ctx, Marker{}, ListBlobsSegmentOptions{
-		Details: BlobListingDetails{Versions: true, Snapshots: true},
-	})
-	c.Assert(lbResp, chk.NotNil)
-	if len(lbResp.Segment.BlobItems) < 2 {
-		c.Fail()
-	}
-
-	_, err = blobURL.Delete(ctx, DeleteSnapshotsOptionInclude, BlobAccessConditions{})
-	lbResp, err = containerURL.ListBlobsFlatSegment(ctx, Marker{}, ListBlobsSegmentOptions{
-		Details: BlobListingDetails{Versions: true, Snapshots: true},
-	})
-	c.Assert(lbResp, chk.NotNil)
-	if len(lbResp.Segment.BlobItems) < 2 {
-		c.Fail()
-	}
-	for _, blob := range lbResp.Segment.BlobItems {
-		c.Assert(blob.Snapshot, chk.Equals, "")
-	}
-}
