+++ azblob/zt_url_block_blob_test.go
@@ -6,7 +6,6 @@ import (
 	"encoding/binary"
 	"fmt"
 	"io/ioutil"
-	"log"
 	"math"
 	"time"
 
@@ -900,148 +899,3 @@ func (s *aztestsSuite) TestBlobPutBlockListModifyBlob(c *chk.C) {
 	c.Assert(resp.UncommittedBlocks, chk.HasLen, 0)
 }
 
-func (s *aztestsSuite) TestCopyBlobFromURLWithSASReturnsVID(c *chk.C) {
-	bsu := getBSU()
-	credential, err := getGenericCredential("")
-	if err != nil {
-		c.Fatal("Invalid credential")
-	}
-	container, _ := createNewContainer(c, bsu)
-	defer delContainer(c, container)
-
-	testSize := 4 * 1024 * 1024 // 4MB
-	r, sourceData := getRandomDataAndReader(testSize)
-	sourceDataMD5Value := md5.Sum(sourceData)
-	ctx := context.Background()
-	srcBlob := container.NewBlockBlobURL(generateBlobName())
-	destBlob := container.NewBlockBlobURL(generateBlobName())
-
-	uploadSrcResp, err := srcBlob.Upload(ctx, r, BlobHTTPHeaders{}, Metadata{}, BlobAccessConditions{})
-	c.Assert(err, chk.IsNil)
-	c.Assert(uploadSrcResp.Response().StatusCode, chk.Equals, 201)
-	c.Assert(uploadSrcResp.Response().Header.Get("x-ms-version-id"), chk.NotNil)
-
-	srcBlobParts := NewBlobURLParts(srcBlob.URL())
-
-	srcBlobParts.SAS, err = BlobSASSignatureValues{
-		Protocol:      SASProtocolHTTPS,                     // Users MUST use HTTPS (not HTTP)
-		ExpiryTime:    time.Now().UTC().Add(48 * time.Hour), // 48-hours before expiration
-		ContainerName: srcBlobParts.ContainerName,
-		BlobName:      srcBlobParts.BlobName,
-		Permissions:   BlobSASPermissions{Read: true}.String(),
-	}.NewSASQueryParameters(credential)
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	srcBlobURLWithSAS := srcBlobParts.URL()
-
-	resp, err := destBlob.CopyFromURL(ctx, srcBlobURLWithSAS, Metadata{"foo": "bar"}, ModifiedAccessConditions{}, BlobAccessConditions{}, sourceDataMD5Value[:])
-	c.Assert(err, chk.IsNil)
-	c.Assert(resp.Response().StatusCode, chk.Equals, 202)
-	c.Assert(resp.Version(), chk.Not(chk.Equals), "")
-	c.Assert(resp.CopyID(), chk.Not(chk.Equals), "")
-	c.Assert(string(resp.CopyStatus()), chk.DeepEquals, "success")
-	c.Assert(resp.VersionID(), chk.NotNil)
-
-	downloadResp, err := destBlob.BlobURL.Download(ctx, 0, CountToEnd, BlobAccessConditions{}, false)
-	c.Assert(err, chk.IsNil)
-	destData, err := ioutil.ReadAll(downloadResp.Body(RetryReaderOptions{}))
-	c.Assert(err, chk.IsNil)
-	c.Assert(destData, chk.DeepEquals, sourceData)
-	c.Assert(downloadResp.Response().Header.Get("x-ms-version-id"), chk.NotNil)
-	c.Assert(len(downloadResp.NewMetadata()), chk.Equals, 1)
-	_, badMD5 := getRandomDataAndReader(16)
-	_, err = destBlob.CopyFromURL(ctx, srcBlobURLWithSAS, Metadata{}, ModifiedAccessConditions{}, BlobAccessConditions{}, badMD5)
-	c.Assert(err, chk.NotNil)
-
-	resp, err = destBlob.CopyFromURL(ctx, srcBlobURLWithSAS, Metadata{}, ModifiedAccessConditions{}, BlobAccessConditions{}, nil)
-	c.Assert(err, chk.IsNil)
-	c.Assert(resp.Response().StatusCode, chk.Equals, 202)
-	c.Assert(resp.XMsContentCrc64(), chk.Not(chk.Equals), "")
-	c.Assert(resp.Response().Header.Get("x-ms-version"), chk.Equals, ServiceVersion)
-	c.Assert(resp.Response().Header.Get("x-ms-version-id"), chk.NotNil)
-}
-
-func (s *aztestsSuite) TestCreateBlockBlobReturnsVID(c *chk.C) {
-	bsu := getBSU()
-	containerURL, _ := createNewContainer(c, bsu)
-	defer delContainer(c, containerURL)
-
-	testSize := 2 * 1024 * 1024 // 1MB
-	r, _ := getRandomDataAndReader(testSize)
-	ctx := context.Background() // Use default Background context
-	blobURL := containerURL.NewBlockBlobURL(generateBlobName())
-
-	// Prepare source blob for copy.
-	uploadResp, err := blobURL.Upload(ctx, r, BlobHTTPHeaders{}, Metadata{}, BlobAccessConditions{})
-	c.Assert(err, chk.IsNil)
-	c.Assert(uploadResp.Response().StatusCode, chk.Equals, 201)
-	c.Assert(uploadResp.rawResponse.Header.Get("x-ms-version"), chk.Equals, ServiceVersion)
-	c.Assert(uploadResp.Response().Header.Get("x-ms-version-id"), chk.NotNil)
-
-	csResp, err := blobURL.CreateSnapshot(ctx, Metadata{}, BlobAccessConditions{})
-	c.Assert(err, chk.IsNil)
-	c.Assert(csResp.Response().StatusCode, chk.Equals, 201)
-	c.Assert(csResp.Response().Header.Get("x-ms-version-id"), chk.NotNil)
-
-	listBlobResp, err := containerURL.ListBlobsFlatSegment(ctx, Marker{}, ListBlobsSegmentOptions{Details: BlobListingDetails{Snapshots: true}})
-	c.Assert(err, chk.IsNil)
-	c.Assert(listBlobResp.rawResponse.Header.Get("x-ms-request-id"), chk.NotNil)
-	if len(listBlobResp.Segment.BlobItems) < 2 {
-		c.Fail()
-	}
-
-	deleteResp, err := blobURL.Delete(ctx, DeleteSnapshotsOptionOnly, BlobAccessConditions{})
-	c.Assert(err, chk.IsNil)
-	c.Assert(deleteResp.Response().StatusCode, chk.Equals, 202)
-	c.Assert(deleteResp.Response().Header.Get("x-ms-version-id"), chk.NotNil)
-
-	listBlobResp, err = containerURL.ListBlobsFlatSegment(ctx, Marker{}, ListBlobsSegmentOptions{Details: BlobListingDetails{Snapshots: true, Versions: true}})
-	c.Assert(err, chk.IsNil)
-	c.Assert(listBlobResp.rawResponse.Header.Get("x-ms-request-id"), chk.NotNil)
-	if len(listBlobResp.Segment.BlobItems) == 0 {
-		c.Fail()
-	}
-	blobs := listBlobResp.Segment.BlobItems
-	c.Assert(blobs[0].Snapshot, chk.Equals, "")
-}
-
-func (s *aztestsSuite) TestPutBlockListReturnsVID(c *chk.C) {
-	blockIDIntToBase64 := func(blockID int) string {
-		binaryBlockID := (&[4]byte{})[:]
-		binary.LittleEndian.PutUint32(binaryBlockID, uint32(blockID))
-		return base64.StdEncoding.EncodeToString(binaryBlockID)
-	}
-	bsu := getBSU()
-	containerURL, _ := createNewContainer(c, bsu)
-	defer delContainer(c, containerURL)
-
-	blobURL := containerURL.NewBlockBlobURL(generateBlobName())
-
-	data := []string{"Azure ", "Storage ", "Block ", "Blob."}
-	base64BlockIDs := make([]string, len(data))
-
-	for index, d := range data {
-		base64BlockIDs[index] = blockIDIntToBase64(index)
-		resp, err := blobURL.StageBlock(ctx, base64BlockIDs[index], strings.NewReader(d), LeaseAccessConditions{}, nil)
-		if err != nil {
-			log.Fatal(err)
-		}
-		c.Assert(resp.Response().StatusCode, chk.Equals, 201)
-		c.Assert(resp.Version(), chk.Not(chk.Equals), "")
-	}
-
-	commitResp, err := blobURL.CommitBlockList(ctx, base64BlockIDs, BlobHTTPHeaders{}, Metadata{}, BlobAccessConditions{})
-	c.Assert(err, chk.IsNil)
-	c.Assert(commitResp.VersionID(), chk.NotNil)
-
-	contentResp, err := blobURL.Download(ctx, 0, CountToEnd, BlobAccessConditions{}, false)
-	c.Assert(err, chk.IsNil)
-	contentData, err := ioutil.ReadAll(contentResp.Body(RetryReaderOptions{}))
-	c.Assert(contentData, chk.DeepEquals, []uint8(strings.Join(data, "")))
-}
-
-func (s *aztestsSuite) TestSyncCopyBlobReturnsVID(c *chk.C) {
-
-}
